# src/docs/swagger.yaml
basePath: /
definitions:
  github_com_gothi_vouchrs_src_delivery_http_response.APIError:
    properties:
      code:
        example: NOT_FOUND
        type: string
      message:
        example: listing not found
        type: string
    type: object
  github_com_gothi_vouchrs_src_delivery_http_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.APIError'
      meta: {}
      success:
        example: true
        type: boolean
    type: object
...
  github_com_gothi_vouchrs_src_internal_domain_entity.Transaction:
    properties:
      buyerAmount:
        type: number
      buyerID:
        type: string
      codeRevealedAt:
        description: when email was sent
        type: string
      completedAt:
        type: string
      createdAt:
        type: string
      id:
        type: string
      listingID:
        type: string
      lockStartedAt:
        type: string
      paidAt:
        type: string
      paymentRef:
        description: PG merchant transaction ID
        type: string
      payoutRef:
        description: Razorpay payout ID
        type: string
      sellerID:
        type: string
      sellerPayout:
        type: number
      status:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.TransactionStatus'
      updatedAt:
        type: string
    type: object
  github_com_gothi_vouchrs_src_internal_domain_entity.TransactionStatus:
    enum:
    - pending
    - paid
    - completed
    - cancelled
    - refunded
    type: string
    x-enum-comments:
      TxnStatusCompleted: buyer confirmed redemption
      TxnStatusPaid: payment confirmed, code sent by email
      TxnStatusPending: lock acquired, waiting for payment
    x-enum-descriptions:
    - lock acquired, waiting for payment
    - payment confirmed, code sent by email
    - buyer confirmed redemption
    - ""
    - ""
    x-enum-varnames:
    - TxnStatusPending
    - TxnStatusPaid
    - TxnStatusCompleted
    - TxnStatusCancelled
    - TxnStatusRefunded
...
  src_delivery_http_handler.initiateBuyResponse:
    properties:
      amount:
        example: 910
        type: number
      lock_expires_at:
        example: "2024-01-01T10:10:00Z"
        type: string
      payment_url:
        example: https://api-preprod.phonepe.com/apis/pg-sandbox/...
        type: string
      transaction_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cardswap.in
    name: CardSwap Support
  description: P2P gift card exchange marketplace. Card codes are AES-256 encrypted
    at rest and **never returned in API responses** — they are delivered to the buyer's
    registered email only.
  title: CardSwap India API
  version: "1.0"
paths:
...
  /api/v1/listings/{id}/buy:
    post:
      description: Runs Gate 2 re-verification, atomically locks the listing for 10
        minutes, creates a pending transaction, and returns a PhonePe payment URL.
        The buyer must complete payment before lock_expires_at or the listing is released.
        **The card code is never returned here** — it is sent to the buyer's email
        after payment succeeds.
      parameters:
      - description: Listing UUID to purchase
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/src_delivery_http_handler.initiateBuyResponse'
              type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "403":
          description: Buyer is banned or trying to buy own listing
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "409":
          description: Listing locked or not available (LISTING_LOCKED / LISTING_NOT_AVAILABLE)
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "422":
          description: Gate 2 failed — card was tampered (CARD_TAMPERED)
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: Initiate purchase (buyer)
      tags:
      - purchase
...
  /api/v1/transactions/{id}:
    get:
      description: Returns a transaction. Only the buyer or seller of that transaction
        can access it.
      parameters:
      - description: Transaction UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.Transaction'
              type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "403":
          description: Not a party to this transaction
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: Get transaction details
      tags:
      - purchase
  /api/v1/transactions/{id}/confirm:
    post:
      description: Called by the buyer after they have successfully used the gift
        card. Marks the transaction as completed and triggers the seller payout.
      parameters:
      - description: Transaction UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: redemption confirmed
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "403":
          description: Not the buyer of this transaction
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: Confirm card redemption (buyer)
      tags:
      - purchase
...
  /api/v1/webhooks/phonepe:
    post:
      consumes:
      - application/json
      description: 'Receives payment SUCCESS/FAILURE events from PhonePe. Verified
        via X-VERIFY checksum header. On success: marks listing sold, emails card
        code to buyer (never in response body), queues payout. On failure: unlocks
        listing.'
      parameters:
      - description: PhonePe SHA256 checksum
        in: header
        name: X-VERIFY
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: acknowledged
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Invalid signature or body
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      summary: PhonePe payment webhook (internal)
      tags:
      - webhooks
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"